src/requirements.txt
fastapi
uvicorn
jinja2
python-multipart
requests
python-magic